feat: 딜러 생성 구현, 딜러는 카드의 이름들과 카드 뭉치의 합을 반환하는 기능 구현
